backend/requirements.txt
langgraph
langchain-google-genai
google-genai

requests
shodan
greynoise
vt-py
python-dotenv
langchain-community
langchain-text-splitters

unstructured

fastapi
uvicorn
sqlmodel
passlib[bcrypt]
python-jose[cryptography]
python-multipart
langchain-chroma
chromadb

detoxify
transformers
torch
safetensors
